feat(scroll-text-slide): remove quote slide for text slide
