--- src/scripts/components/stories/story/blocks/story-eei/modules/quote-slide/quote-slide.module.css
@@ -1,28 +1,0 @@
-.quoteWrapper {
-  display: grid;
-  place-items: center;
-  width: min(90%, 600px);
-  margin: 0 auto;
-
-  .quote {
-    text-shadow: 0 2px 1px rgba(0, 0, 0, 0.3);
-    height: fit-content;
-    top: 25%;
-
-    div {
-      padding-inline: 1.5rem;
-      border-left: 4px solid var(--Background-ESA-Teal, #00b398);
-      margin-bottom: 25vh;
-
-      p {
-        line-height: 150%; /* 30px */
-        font-size: 1.25rem;
-        font-weight: 700;
-        margin-bottom: 1rem;
-      }
-      span {
-        color: var(--font-55-esa-teal, #98dbce);
-      }
-    }
-  }
-}
