+++ storage/stories/story-eei/story-eei-de.json
@@ -529,7 +529,7 @@
         }
       ],
       "lengthFactor": 6,
-      "type": "quoteSlide",
+      "type": "scrollTextSlide",
       "content": {
         "text": "ESA’s Climate Change Initiative project MOTECUSOMA (Monitoring the Energy Cycle for a Better Understanding of Climate Change) is calculating new estimates of Earth’s energy imbalance and surface air temperature. Scientific analyses using these datasets will help climate modelers forecast future changes with more accuracy.",
         "author": "- The MOTECUSOMA Project Team"
